+++ build.sbt
@@ -15,7 +15,7 @@ lazy val http4sDomVersion         = "0.2.9"
 lazy val lucumaBCVersion          = "0.4.0"
 lazy val lucumaCoreVersion        = "0.82.0"
 lazy val lucumaPrimeStylesVersion = "0.2.9"
-lazy val lucumaReactVersion       = "0.42.0"
+lazy val lucumaReactVersion       = "0.42.0-6-0b42e4e-20230825T215554Z-SNAPSHOT"
 lazy val lucumaRefinedVersion     = "0.1.2"
 lazy val lucumaSchemasVersion     = "0.58.0"
 lazy val lucumaSsoVersion         = "0.6.4"
@@ -56,7 +56,7 @@ lazy val demo =
   project
     .in(file("modules/demo"))
     .enablePlugins(ScalaJSPlugin, NoPublishPlugin)
-    .dependsOn(ui)
+    .dependsOn(ui, css)
     .settings(
       Compile / scalaJSLinkerConfig ~= { _.withModuleKind(ModuleKind.ESModule) },
       Compile / fastLinkJS / scalaJSLinkerConfig ~= (_.withModuleSplitStyle(
