+++ modules/ui/src/main/resources/lucuma-css/lucuma-ui-sequence.scss
@@ -13,3 +13,23 @@
 .lucuma-sequece-step-guided {
   color: lightgreen;
 }
+
+.lucuma-sequece-step-type-object {
+  color: var(--green-500);
+}
+
+.lucuma-sequece-step-type-arc {
+  color: var(--purple-600);
+}
+
+.lucuma-sequece-step-type-flat {
+  color: var(--gray-600);
+}
+
+.lucuma-sequece-step-type-bias {
+  color: var(--teal-400);
+}
+
+.lucuma-sequece-step-type-dark {
+  color: var(--gray-900);
+}
