+++ modules/ui/src/main/scala/lucuma/ui/sequence/SequenceStyles.scala
@@ -9,3 +9,10 @@ object SequenceStyles:
   val RightAligned: Css = Css("lucuma-sequece-right-aligned")
   val LeftAligned: Css  = Css("lucuma-sequece-left-aligned")
   val StepGuided: Css   = Css("lucuma-sequece-step-guided")
+
+  object StepType:
+    val Bias: Css   = Css("lucuma-sequece-step-type-bias")
+    val Dark: Css   = Css("lucuma-sequece-step-type-dark")
+    val Arc: Css    = Css("lucuma-sequece-step-type-arc")
+    val Flat: Css   = Css("lucuma-sequece-step-type-flat")
+    val Object: Css = Css("lucuma-sequece-step-type-object")
